+++ CHANGELOG.md
@@ -12,10 +12,11 @@
   * Value of _pluginName_ variable depends on build-time variable _NetworkName_. Default value - ethereum_main_net
 ### Changed
 * Added _**HDWALLET_CHAIN_ID**_ environment variable to helm-chart description
+* Removed proto description of evm transactions. Now used native go-ethereum binary marshaling/unmarshaling binary data flow 
 
 ## [v0.0.32] 09.06.2024
 ### Added
-* Added proto description of Ethereum transaction types:
+* ~~Added proto description of Ethereum transaction types~~:
   * LegacyTx
   * DynamicFeeTx
   * AccessListTx